# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VisiteCRM API - SaaS CRM for tourism agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all platform users (superadmin only)
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserItem"
...
  /admin/audit-logs:
    get:
      operationId: listAdminAuditLogs
      tags: [admin]
      summary: List all audit logs (superadmin only)
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogWithTenant"
...
  /tenants/{id}/details:
    get:
      operationId: getTenantDetails
      tags: [tenants]
      summary: Get detailed tenant info with counts (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetails"
  /tenants/{id}/users:
    get:
      operationId: listTenantUsers
      tags: [tenants]
      summary: List users of a tenant (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /tenants/{id}/suspend:
    post:
      operationId: suspendTenant
      tags: [tenants]
      summary: Suspend a tenant (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}/activate:
    post:
      operationId: activateTenant
      tags: [tenants]
      summary: Activate a tenant (superadmin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ─── DASHBOARD ───────────────────────────────────────────────────
...
  /automation-actions:
    get:
      operationId: listAutomationActions
      summary: List automation actions
      tags: [AutomationExtras]
      responses:
        "200":
          description: List of automation actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationAction"
    post:
      operationId: createAutomationAction
      summary: Create an automation action
      tags: [AutomationExtras]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationActionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationAction"
  /automation-actions/{id}:
    patch:
      operationId: updateAutomationAction
      summary: Update an automation action
      tags: [AutomationExtras]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationActionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationAction"
    delete:
      operationId: deleteAutomationAction
      summary: Delete an automation action
      tags: [AutomationExtras]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Role:
      type: string
      enum: [superadmin, agencia, gerente, vendedor, suporte, cliente]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        tenantId:
          type: ["string", "null"]
        referralCode:
          type: string
        referralBalance:
          type: number
        commissionType:
          type: string
        commissionRate:
          type: number
        commissionFixed:
          type: number
        monthlyGoal:
          type: ["number", "null"]
        createdAt:
          type: string
        tenant:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
            logoUrl:
              type: ["string", "null"]
            primaryColor:
              type: ["string", "null"]
            secondaryColor:
              type: ["string", "null"]
            status:
              type: string
            planId:
              type: string
            website:
              type: ["string", "null"]
          required: [id, name, slug, status, planId]
      required: [id, clerkId, name, email, role, isActive, referralCode, referralBalance, createdAt]
...
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        email:
          type: string
        whatsapp:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        planId:
          type: string
        status:
          type: string
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        reservationPrefix:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, slug, email, planId, status, createdAt]
...
    AutomationAction:
      type: object
      properties:
        id:
          type: string
        automationId:
          type: string
        tenantId:
          type: string
        type:
          type: string
        config:
          type: object
        order:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, automationId, tenantId, type, config, order, isActive, createdAt]
    CreateAutomationActionBody:
      type: object
      properties:
        automationId:
          type: string
        type:
          type: string
        config:
          type: object
        order:
          type: integer
        isActive:
          type: boolean
      required: [automationId, type]
...
    TenantDetails:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        email:
          type: string
        whatsapp:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        planId:
          type: string
        status:
          type: string
        cnpj:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        userCount:
          type: integer
        clientCount:
          type: integer
        tripCount:
          type: integer
        reservationCount:
          type: integer
        planMaxUsers:
          type: ["integer", "null"]
        planMaxClients:
          type: ["integer", "null"]
        planMaxTrips:
          type: ["integer", "null"]
        maxUsersOverride:
          type: ["integer", "null"]
        maxClientsOverride:
          type: ["integer", "null"]
        maxTripsOverride:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, slug, email, planId, status, userCount, clientCount, tripCount, reservationCount, createdAt, updatedAt]

    # ─── ADMIN USER ─────────────────────────────────────────────────
    AdminUserItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        tenantId:
          type: ["string", "null"]
        tenantName:
          type: ["string", "null"]
        tenantStatus:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, isActive, createdAt]

    # ─── AUDIT LOG WITH TENANT ──────────────────────────────────────
    AuditLogWithTenant:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: ["string", "null"]
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        before:
          type: object
          nullable: true
        after:
          type: object
          nullable: true
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
        tenantName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
      required: [id, tenantId, action, entityType, entityId, createdAt]

    # ─── SUBSCRIPTION ───────────────────────────────────────────────
...